backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0

# Pydantic v2 stack
pydantic==2.11.1
pydantic-settings==2.5.2

python-multipart==0.0.9
numpy==2.1.2
pillow==11.0.0
PyPDF2==3.0.1
pdfminer.six==20231228
ultralytics
